Fix by comments.
