+++ huntflow_api_client/models/response/applicants.py
@@ -340,14 +340,3 @@ class ApplicantLogItem(BaseModel):
 
 class ApplicantLogResponse(PaginatedResponse):
     items: List[ApplicantLogItem] = Field(..., description="List of applicant's logs")
-
-
-class ApplicantSource(BaseModel):
-    id: int = Field(..., description="Applicant source ID")
-    foreign: Optional[str] = Field(None, description="Applicant source foreign")
-    name: str = Field(..., description="Applicant source name")
-    type: str = Field(..., description="Applicant source type")
-
-
-class ApplicantSourcesResponse(BaseModel):
-    items: List[ApplicantSource] = Field(..., description="List of applicant's sources")
